(maint) add grouped jobs concept and count interface
This adds the concept of jobs within the same group.  Jobs that are grouped
can be counted or stopped together.  This is helpful when managing jobs associated
with a given trapperkeeper service.

Upon service stop, the jobs set is now emptied.  Previously the jobs were stopped
but the set maintained the entries in the set.

`after` calls are now given a unique identifier and removed from the set when they
complete.