(MAINT) Avoid a null pointer exception during an early shutdown
Previously, if a startup error were encountered before the
scheduler service's init function was called, a call to the
stop function could throw/log a NullPointerException when trying
to access state that had not been set on the context during the
init call.

With this commit, the scheduler service just avoids the code
which would access the context state that doesn't exist,
avoiding the appearance of the NullPointerException.