(PE-23981) move from at-at to quartz
This moves the internal implementation from the no-longer maintained
at-at library to the heavily maintained quartz library.

It reimplements the group-id concept using the underlying group concept
built into quartz, and adds interfaces for retrieving the existing job
identifiers.

It is a functionally equivalent implementation.

In addition the library has been updated to require java 8.