(PE-25389) add support for iterval and interval-after
This adds two new interface functions for the service: `interval` and
`interval-after`.  An interval job that runs at a given cadence.

This also adds the ability to configure the number of threads used
by the quartz scheduler, and defaults to 10 if it isn't specified.

This also updates the README to describe the interface.