(maint) Use unique scheduler rather than default
File sync integration tests require running the client service and the
server service as separate TK apps in the same process. By grabbing the
default scheduler we end up sharing one schduler between tk apps w/in
the process. In the tests, as we shutdown the server (to test that the
client can continue w/o updates and can reconnect when the server comes
back up), the client dies since its scheduler is shutdown with the
server service.

This may have some use in some PE deployment scenarios (or jar-jar
scenarios), but mainly I've had a terrible time coming up with as
valuable tests for file-sync w/o this feature.