(maint) Help clj find correct StdScheduleFactory constructor
Clojure finds classes and constructors using Java's Class.forName with
an explicit classloader. Which classloader that is however is determined
dynamically within clojure.lang.RT.baseLoader().

The single arity constructor is overloaded for this factory, and without
a compiler hint it appears that on reload the wrong classloader is
requested to load this factory.